This merges the bugfix from main (v4.1.1) that corrects the instanceId datatype from int to long in DeleteReleaseFromCollectionFolder method.

Changes:
- Updated IDiscogsApiClient.DeleteReleaseFromCollectionFolder to use long instanceId
- Updated CollectionFolderReleasesTests to use long for instanceId parameters
- Maintained modernization branch style (ArgumentException.ThrowIf* and ConfigureAwait(false))

+++ src/DiscogsApiClient.Tests/Collection/CollectionFolderReleasesTests.cs
@@ -310,7 +310,7 @@ public sealed class CollectionFolderReleasesTests(DiscogsApiClientFixture fixtur
     [Test]
     [Arguments(-1)]
     [Arguments(0)]
-    public async Task DeleteReleaseFromCollectionFolder_ShouldThrowArgumentOutOfRangeException_WhenInstanceIdIsInvalid(int instanceId, CancellationToken cancellationToken)
+    public async Task DeleteReleaseFromCollectionFolder_ShouldThrowArgumentOutOfRangeException_WhenInstanceIdIsInvalid(long instanceId, CancellationToken cancellationToken)
     {
         var username = "DamIDhagor";
         var folderId = 999;
@@ -326,7 +326,7 @@ public sealed class CollectionFolderReleasesTests(DiscogsApiClientFixture fixtur
         var username = "DamIDhagor";
         var folderId = 999;
         var releaseId = 5134861;
-        var instanceId = int.MaxValue;
+        var instanceId = long.MaxValue;
 
         await Assert.That(async () => await _apiClient.DeleteReleaseFromCollectionFolder(username, folderId, releaseId, instanceId, cancellationToken))
             .Throws<ResourceNotFoundDiscogsException>();
